requirements.txt
howlongtobeatpy==1.0.16
requests==2.32.3
python-dotenv==1.0.1
halo